AI_Evaluation/requirements.txt
fastapi
uvicorn
google-cloud-vision
sentence-transformers
scikit-learn
python-multipart
PyMuPDF
numpy
google-generativeai
pandas